trendy_backend/trendy_backend/requirements-minimal.txt
fastapi==0.68.0
uvicorn[standard]==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-dotenv==0.19.0
stripe==2.60.0
python-multipart==0.0.5
websockets==10.4
aioredis==2.0.0
psycopg2-binary==2.9.1